requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

# LangGraph + LLM
langgraph==0.2.28
langchain==0.3.0
langchain-openai==0.2.0
openai==1.47.0

# Database
supabase==2.7.4
asyncpg==0.29.0

# Telephony + Voice
twilio==9.3.2
retell-sdk==4.1.0
elevenlabs==1.9.0

# Payments
stripe==10.12.0

# HTTP client
httpx==0.27.2
aiohttp==3.10.5

# Logging
structlog==24.4.0

# Environment
python-dotenv==1.0.1

# Utilities
tenacity==8.3.0
python-jose==3.3.0

google-api-python-client==2.139.0
google-auth==2.35.0
tzdata==2024.2